requirements.txt
aiohttp>=3.9.0
textual>=0.47.0
python-dotenv>=1.0.0
